fix(web): 🐛 Harden visual story publication validation
